# services/reasoner/contracts/openapi.yaml
openapi: 3.1.0

info:
  title: Sherlock Reasoning Service
  version: 0.1.0
  description: |
    LangGraph reasoning engine with pgvector memory (PostgreSQL + pgvector).
    Accepts reasoning requests over HTTP (sync). For inter-service messaging
    see `asyncapi.yaml` — NATS request-reply and Pulsar durable async.
  contact:
    name: A.R.C. Platform
    url: https://github.com/arc-framework/arc-platform

servers:
  - url: http://localhost:8083
    description: Local development

tags:
  - name: reasoning
    description: Core reasoning endpoints
  - name: health
    description: Service health and readiness probes
  - name: chat
    description: OpenAI-compatible chat completion API (v1)
  - name: models
    description: Model registry

x-tagGroups:
  - name: Sherlock Core
    tags:
      - reasoning
      - health
  - name: Sherlock Chat API (v1)
    tags:
      - chat
      - models

paths:
  /chat:
    post:
      operationId: chat
      tags: [reasoning]
      summary: Submit a reasoning request
      description: |
        Synchronous reasoning request. Sherlock retrieves semantic context from pgvector,
        generates a response via Ollama (LangGraph graph), and persists both turns to
        pgvector and PostgreSQL before returning.

        Latency is bounded by LLM inference time (default model: `mistral:7b`).
        For long-running tasks prefer NATS request-reply or Pulsar durable async.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              basic:
                summary: First message
                value:
                  user_id: "user-123"
                  text: "What is the capital of France?"
              follow_up:
                summary: Follow-up using conversation history
                value:
                  user_id: "user-123"
                  text: "What language do they speak there?"
      responses:
        '200':
          description: Reasoning completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                success:
                  value:
                    user_id: "user-123"
                    text: "The capital of France is Paris."
                    latency_ms: 1250
        '422':
          description: Validation error — missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                missing_user_id:
                  value:
                    detail:
                      - loc: ["body", "user_id"]
                        msg: "Field required"
                        type: "missing"
        '500':
          description: Internal error during graph execution
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                llm_error:
                  value:
                    detail: "LLM unavailable after 3 retries"
        '503':
          description: Service not ready — startup incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_ready:
                  value:
                    detail: "Service not ready"

  /health:
    get:
      operationId: health
      tags: [health]
      summary: Shallow liveness check
      description: |
        Always returns 200 if the process is alive and the NATS connection
        was established during startup. Does NOT probe PostgreSQL.
        Use `/health/deep` for readiness checks.
      responses:
        '200':
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                alive:
                  value:
                    status: "ok"
                    version: "0.1.0"

  /health/deep:
    get:
      operationId: health_deep
      tags: [health]
      summary: Deep readiness check
      description: |
        Probes all dependencies — PostgreSQL and NATS — and returns
        per-component status. Returns 503 if any dependency is degraded.
        Used by Docker healthcheck and the CLI `make dev-health` target.
      responses:
        '200':
          description: All dependencies healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepHealthResponse'
              examples:
                all_healthy:
                  value:
                    status: "ok"
                    version: "0.1.0"
                    components:
                      postgres: true
                      nats: true
        '503':
          description: One or more dependencies are unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepHealthResponse'
              examples:
                postgres_down:
                  value:
                    status: "degraded"
                    version: "0.1.0"
                    components:
                      postgres: false
                      nats: true

  /v1/models:
    get:
      operationId: list_models
      tags: [models]
      summary: List available models
      description: |
        Returns the single configured LLM model in OpenAI-compatible format.
        The model ID is controlled by `SHERLOCK_LLM_MODEL`.
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
              examples:
                default:
                  value:
                    object: list
                    data:
                      - id: "mistralai/mistral-7b-instruct-v0.3"
                        object: model
                        owned_by: arc-platform

  /v1/chat/completions:
    post:
      operationId: v1_chat_completions
      tags: [chat]
      summary: OpenAI-compatible chat completion
      description: |
        Synchronous or streaming chat completion in OpenAI wire format.
        Set `stream: true` for Server-Sent Events (SSE) token streaming.

        **Model validation**: only the configured `SHERLOCK_LLM_MODEL` is accepted.
        **User ID**: `user` field optional — derived via UUID v5 when absent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              sync:
                summary: Synchronous completion
                value:
                  model: "mistralai/mistral-7b-instruct-v0.3"
                  messages:
                    - role: user
                      content: "What is 2+2?"
                  stream: false
                  user: "user-123"
              stream:
                summary: Streaming completion (SSE)
                value:
                  model: "mistralai/mistral-7b-instruct-v0.3"
                  messages:
                    - role: user
                      content: "Write a short poem"
                  stream: true
      responses:
        '200':
          description: |
            Sync: `ChatCompletionResponse` JSON.
            Stream: `text/event-stream` SSE with `ChatCompletionChunk` events, terminated by `data: [DONE]`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: |
                  Server-Sent Events stream. Each line is `data: {ChatCompletionChunk JSON}`.
                  Stream ends with `data: [DONE]`.
              examples:
                stream_events:
                  value: "data: {\"id\":\"chatcmpl-abc\",\"object\":\"chat.completion.chunk\",...}\n\ndata: [DONE]\n\n"
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1ErrorResponse'
              examples:
                model_not_found:
                  value:
                    error:
                      type: invalid_request_error
                      code: model_not_found
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/responses:
    post:
      operationId: v1_responses
      tags: [chat]
      summary: Responses API (OpenAI-compatible)
      description: |
        Simplified responses endpoint — accepts string or structured input,
        returns a `ResponsesResponse` with `output[].content[].text`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
            examples:
              string_input:
                summary: String input
                value:
                  model: "mistralai/mistral-7b-instruct-v0.3"
                  input: "What is 2+2?"
                  user: "user-123"
      responses:
        '200':
          description: Response completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsesResponse'
        '404':
          description: Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1ErrorResponse'
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - user_id
        - text
      properties:
        user_id:
          type: string
          description: Unique identifier for the user session. Used to scope memory and context retrieval.
          example: "user-123"
          minLength: 1
        text:
          type: string
          description: The user's message to reason about.
          example: "What is the capital of France?"
          minLength: 1

    ChatResponse:
      type: object
      required:
        - user_id
        - text
        - latency_ms
      properties:
        user_id:
          type: string
          description: Echo of the request user_id.
          example: "user-123"
        text:
          type: string
          description: Sherlock's reasoned response.
          example: "The capital of France is Paris."
        latency_ms:
          type: integer
          description: Total end-to-end processing time in milliseconds.
          example: 1250
          minimum: 0

    HealthResponse:
      type: object
      required:
        - status
        - version
      properties:
        status:
          type: string
          enum: [ok]
          example: "ok"
        version:
          type: string
          description: Service version from settings.
          example: "0.1.0"

    DeepHealthResponse:
      type: object
      required:
        - status
        - version
        - components
      properties:
        status:
          type: string
          enum: [ok, degraded]
          example: "ok"
        version:
          type: string
          example: "0.1.0"
        components:
          type: object
          required:
            - postgres
            - nats
          properties:
            postgres:
              type: boolean
              description: PostgreSQL (Oracle) reachable and schema initialized.
              example: true
            nats:
              type: boolean
              description: NATS (Flash) connection established.
              example: true

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "Service not ready"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            required:
              - loc
              - msg
              - type
            properties:
              loc:
                type: array
                items:
                  type: string
                example: ["body", "user_id"]
              msg:
                type: string
                example: "Field required"
              type:
                type: string
                example: "missing"

    # ── v1 Chat API schemas ────────────────────────────────────────

    ChatMessage:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
          nullable: true
        name:
          type: string
          nullable: true

    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          example: "mistralai/mistral-7b-instruct-v0.3"
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 2
          nullable: true
        max_tokens:
          type: integer
          minimum: 1
          nullable: true
        top_p:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        user:
          type: string
          nullable: true
          description: Optional user identifier. Derived via UUID v5 when absent.

    UsageInfo:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer

    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          example: "chatcmpl-abc123"
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          description: Unix timestamp
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index:
                type: integer
              message:
                $ref: '#/components/schemas/ChatMessage'
              finish_reason:
                type: string
                nullable: true
                enum: [stop, length, content_filter, tool_calls]
        usage:
          $ref: '#/components/schemas/UsageInfo'

    ChatCompletionChunk:
      type: object
      required: [id, object, created, model, choices]
      description: SSE streaming chunk
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion.chunk]
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [index, delta]
            properties:
              index:
                type: integer
              delta:
                type: object
                properties:
                  role:
                    type: string
                    nullable: true
                  content:
                    type: string
                    nullable: true
              finish_reason:
                type: string
                nullable: true

    ResponsesRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                required: [role, content]
                properties:
                  role:
                    type: string
                    enum: [user, assistant, system]
                  content:
                    type: string
        instructions:
          type: string
          nullable: true
        user:
          type: string
          nullable: true

    ResponsesResponse:
      type: object
      required: [id, object, created_at, model, output, usage]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [response]
        created_at:
          type: integer
        model:
          type: string
        output:
          type: array
          items:
            type: object
            required: [type, role, content]
            properties:
              type:
                type: string
                enum: [message]
              role:
                type: string
                enum: [assistant]
              content:
                type: array
                items:
                  type: object
                  required: [type, text]
                  properties:
                    type:
                      type: string
                      enum: [output_text]
                    text:
                      type: string
        usage:
          type: object
          required: [input_tokens, output_tokens, total_tokens]
          properties:
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        status:
          type: string
          enum: [completed, failed, in_progress]
          default: completed

    ModelObject:
      type: object
      required: [id, object, owned_by]
      properties:
        id:
          type: string
          example: "mistralai/mistral-7b-instruct-v0.3"
        object:
          type: string
          enum: [model]
        created:
          type: integer
        owned_by:
          type: string
          example: arc-platform

    ModelList:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: '#/components/schemas/ModelObject'

    V1ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type]
          properties:
            type:
              type: string
              example: invalid_request_error
            code:
              type: string
              example: model_not_found
            message:
              type: string
              nullable: true